--- cmd/agent/README.md
@@ -1,3 +1,0 @@
-# cmd/agent
-
-В данной директории будет содержаться код Агента, который скомпилируется в бинарное приложение
