--- cmd/server/README.md
@@ -1,3 +1,0 @@
-# cmd/agent
-
-В данной директории будет содержаться код Сервера, который скомпилируется в бинарное приложение
