+++ internal/server/README.md
@@ -1,0 +1,3 @@
+# cmd/agent
+
+В данной директории будет содержаться код Сервера, который скомпилируется в бинарное приложение
